pyproject.toml
[project]
name = "seizu"
version = "2024.12.0"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "httpx>=0.28.1",
    "litestar[brotli,standard]>=2.13.0",
    "msgspec>=0.19.0",
    "polars>=1.18.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "datamodel-code-generator[http]>=0.26.4",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "time-machine>=2.16.0",
]

[tool.datamodel-codegen]
url = "https://api.ynab.com/papi/open_api_spec.yaml"
input-file-type = "openapi"
output = "app/core/models.py"
output-model-type = "msgspec.Struct"
use-annotated = true
use-standard-collections = true
use-union-operator = true
capitalise-enum-members = true
disable-timestamp = true
keyword-only = true
reuse-model = true
target-python-version = "3.13"

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
